refactor: cleanup and reorder schemas and routes
